refactored file positions
