+++ Packages/com.nickmaltbie.openkcc.cinemachine/OpenKCC.cinemachine/CameraControls.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c24cf7b1f75e864d886d29d88f2baf1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
