+++ Packages/com.nickmaltbie.openkcc.cinemachine/OpenKCC.cinemachine/CameraControls/CameraPlayerHeading.cs
@@ -1,0 +1,45 @@
+﻿// Copyright (C) 2023 Nicholas Maltbie
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+// associated documentation files (the "Software"), to deal in the Software without restriction,
+// including without limitation the rights to use, copy, modify, merge, publish, distribute,
+// sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all copies or
+// substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
+// BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
+// CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+// ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using nickmaltbie.OpenKCC.CameraControls;
+using UnityEngine;
+
+namespace nickmaltbie.OpenKCC.cinemachine.CameraControls
+{
+    /// <summary>
+    /// Get player heading based on current main camera rotation.
+    /// </summary>
+    public class CameraPlayerHeading : MonoBehaviour, ICameraControls
+    {
+        public Quaternion PlayerHeading
+        {
+            get
+            {
+                // Check if there is a main camera
+                if (Camera.main != null)
+                {
+                    // Use the horizontal component of the camera's view
+                    // to get the player heading.
+                    return Quaternion.Euler(0, Camera.main.transform.rotation.eulerAngles.y, 0);
+                }
+
+                return Quaternion.identity;
+            }
+        }
+    }
+}
