+++ Packages/com.nickmaltbie.openkcc.cinemachine/OpenKCC.cinemachine/CameraControls/CameraSelector.cs
@@ -3,7 +3,7 @@ using Cinemachine;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace nickmaltbie.OpenKCC.cinemachine
+namespace nickmaltbie.OpenKCC.cinemachine.CameraControls
 {
     public class CameraSelector : MonoBehaviour
     {
