+++ Packages/com.nickmaltbie.openkcc.cinemachine/OpenKCC.cinemachine/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5a105d4bfeb965419129b4ecc3c8cac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
