+++ Packages/com.nickmaltbie.openkcc.cinemachine/OpenKCC.cinemachine/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e00a532ad77013247af56f3bb1fc7283
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
